requirements.txt
pandas==1.5.2
plotly==5.11.0
streamlit==1.16.0
beautifulsoup4==4.9.3
ipykernel==5.3.4
ipython==7.19.0
matplotlib==3.3.4
numpy==1.20.3
Pillow==8.1.0
pyclustertend==1.4.9
python-dotenv==0.15.0
requests==2.25.0
scikit-learn==0.24.1
scipy==1.6.0
seaborn==0.11.1
xlrd == 1.2.0
nltk==3.8
openpyxl==3.0.10
spacy>=3.0.0,<4.0.0
https://github.com/explosion/spacy-models/releases/download/es_core_news_sm-3.4.0/es_core_news_sm-3.4.0.tar.gz#egg=es_core_news_sm